Deaths registered weekly in England and Wales, provisional

Provisional counts of the number of deaths registered in England and Wales, by age, sex and region, in the latest weeks for which data are available.

data(ONSMortality)

Format

Data frame with five columns

Category

Data is released in groups: Totals, by age groups, males by age, females by age, All respiratory diseases (ICD-10 J00-J99) ICD-10 v 2001 and by Region

Counts

Counts of deaths in whole numbers

Date

Provisional deaths are released on the Friday of every week.

WeekNo

Following the format of the ONS spreadsheets each week is numbered sequentially within the year.

Source

Collected by Zoë Turner zoe.turner2@nottshc.nhs.uk, Apr-2020 from https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/deaths/datasets/weeklyprovisionalfiguresondeathsregisteredinenglandandwales

Details

Source and licence acknowledgement This data has been made available through Office of National Statistics under the Open Government Licence http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/

Examples

data(ONSMortality)

library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

library(tidyr)

wideForm <- ONSMortality %>%
 select(-WeekNo) %>%
 pivot_wider(names_from = Date,
             values_from = Counts
 )

#> Warning: Values in `Counts` are not uniquely identified; output will contain list-cols.
#> * Use `values_fn = list(Counts = list)` to suppress this warning.
#> * Use `values_fn = list(Counts = length)` to identify where the duplicates arise
#> * Use `values_fn = list(Counts = summary_fun)` to summarise duplicates