Synthetic National Early Warning Scores Data

Synthetic NEWS data to show as the results of the NHSR_synpop package. These datasets have been synthetically generated by this package to be utilised in the NHSRDatasets package.

Usage

data(synthetic_news_data)

Format

Tibble with twelve columns

male

character string containing gender code

age

age of patient

NEWS

National Early Warning Score (NEWS)

syst

Systolic BP - Systolic BP result

dias

Diastolic Blood Pressure - result on NEWS scale

temp

Temperature of patient

pulse

Pulse of the patient

resp

Level of response from the patient

sat

SATS(Oxygen Saturation Levels) of the patient

sup

Suppressed Oxygen score

alert

Level of alertness of patient

died

Indicator to monitor patient death

Source

Generated by Dr. Muhammed Faisal and created by Gary Hutson, Mar-2021

Examples

library(dplyr)

data("synthetic_news_data")

synthetic_news_data |>
  glimpse()
#> Rows: 1,000
#> Columns: 12
#> $ male  <int> 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 1, 1, 0, 1, 0, 0, 1…
#> $ age   <int> 68, 94, 85, 44, 77, 58, 25, 69, 91, 70, 87, 93, 61, 75, 97, 80, …
#> $ NEWS  <int> 3, 1, 0, 0, 1, 1, 4, 0, 1, 1, 7, 2, 5, 1, 1, 3, 1, 5, 0, 2, 1, 2…
#> $ syst  <int> 150, 145, 169, 154, 122, 146, 65, 116, 162, 132, 110, 166, 123, …
#> $ dias  <int> 98, 67, 69, 106, 67, 106, 42, 56, 72, 96, 85, 90, 78, 80, 72, 81…
#> $ temp  <dbl> 36.8, 35.0, 36.2, 36.9, 36.4, 35.3, 35.6, 37.2, 35.5, 35.3, 37.0…
#> $ pulse <int> 78, 62, 54, 80, 62, 73, 72, 90, 60, 67, 95, 87, 93, 65, 89, 145,…
#> $ resp  <int> 26, 18, 18, 17, 20, 20, 12, 16, 16, 16, 24, 16, 26, 12, 16, 16, …
#> $ sat   <int> 96, 96, 96, 96, 95, 98, 99, 97, 99, 97, 87, 95, 96, 96, 98, 99, …
#> $ sup   <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0…
#> $ alert <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…
#> $ died  <int> 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0…