Deaths registered weekly in England and Wales, provisional

Provisional counts of the number of deaths registered in England and Wales, by age, sex and region, in the latest weeks for which data are available.

data(ons_mortality)

Format

Data frame with five columns

category_1

character, containing the names of the groups for counts, e.g. "Total deaths", "all ages".

category_2

character, subcategory of names of groups where necessary, e.g. details of region: "East", details of age bands "15-44".

counts

numeric, numbers of deaths in whole numbers and average numbers with decimal points. To retain the integrity of the format this column data is left as character.

date

date, format is yyyy-mm-dd; all dates are a Friday.

week_no

integer, each week in a year is numbered sequentially.

Source

Collected by Zoë Turner zoe.turner2@nottshc.nhs.uk, Apr-2020 from https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/deaths/datasets/weeklyprovisionalfiguresondeathsregisteredinenglandandwales

Details

Source and licence acknowledgement This data has been made available through Office of National Statistics under the Open Government Licence http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/

Examples

data(ons_mortality)

library(dplyr)
library(tidyr)

wideForm <- ons_mortality %>%
 select(-week_no) %>%
 pivot_wider(names_from = date,
             values_from = counts
 )